REF: move more methods into utils
